fix(home): unify activity collector implementations (#227)
